resolve: previous conversations
